+++ Modules/MetalModule/Sources/MetalModule/Camera/SnapshotProvider/Entities/CameraFollowSnapshotDependency.swift
@@ -1,0 +1,13 @@
+//
+//  CameraFollowSnapshotDependency.swift
+//  MetalModule
+//
+//  Created by max on 31.05.2026.
+//
+
+struct CameraFollowSnapshotDependency: Equatable {
+    let planetName: String
+    let worldPosition: SIMD3<Float>?
+    let framingRadius: Float?
+    let hasActiveTransition: Bool
+}
