+++ Modules/MetalModule/Sources/MetalModule/Camera/SnapshotProvider/Entities/CameraFrameModeState.swift
@@ -1,0 +1,18 @@
+//
+//  CameraFrameModeState.swift
+//  MetalModule
+//
+//  Created by max on 31.05.2026.
+//
+
+struct CameraFrameModeState: Equatable {
+    let navigationControlsCamera: Bool
+    let navigation: CameraNavigationSnapshotDependency?
+    let transferPreviewActive: Bool
+    let transfer: CameraTransferSnapshotDependency?
+
+    var hasActiveExternalCameraMotion: Bool {
+        navigation?.hasActiveTransition == true ||
+        transfer?.hasActiveTransition == true
+    }
+}
