+++ Modules/MetalModule/Sources/MetalModule/Camera/SnapshotProvider/Entities/CameraProjectionParameters.swift
@@ -1,0 +1,11 @@
+//
+//  CameraProjectionParameters.swift
+//  MetalModule
+//
+//  Created by max on 31.05.2026.
+//
+
+struct CameraProjectionParameters: Equatable {
+    let nearPlane: Float
+    let farPlane: Float
+}
