+++ Modules/MetalModule/Sources/MetalModule/Camera/SnapshotProvider/Entities/CameraSnapshotDependencies.swift
@@ -1,0 +1,18 @@
+//
+//  CameraSnapshotDependencies.swift
+//  MetalModule
+//
+//  Created by max on 31.05.2026.
+//
+
+import CoreFoundation
+
+struct CameraSnapshotDependencies: Equatable {
+    let followedObject: CameraFollowSnapshotDependency?
+    let navigation: CameraNavigationSnapshotDependency?
+    let transfer: CameraTransferSnapshotDependency?
+    let activeCameraMotionRevision: Int
+    let sceneFrameID: UInt64?
+    let viewportSize: CGSize
+    let projection: CameraProjectionParameters
+}
