+++ Modules/MetalModule/Sources/MetalModule/Camera/SnapshotProvider/Entities/CameraTransferSnapshotDependency.swift
@@ -1,0 +1,11 @@
+//
+//  CameraTransferSnapshotDependency.swift
+//  MetalModule
+//
+//  Created by max on 31.05.2026.
+//
+
+struct CameraTransferSnapshotDependency: Equatable {
+    let destinationName: String?
+    let hasActiveTransition: Bool
+}
